Pass template dir options to plugins
