+++ src/TemplateConfig.js
@@ -7,6 +7,7 @@ const EleventyBaseError = require("./EleventyBaseError");
 const eleventyConfig = require("./EleventyConfig");
 const debug = require("debug")("Eleventy:TemplateConfig");
 const deleteRequireCache = require("./Util/DeleteRequireCache");
+const aggregateBench = require("./BenchmarkManager").get("Aggregate");
 
 class EleventyConfigError extends EleventyBaseError {}
 
@@ -22,7 +23,7 @@ class TemplateConfig {
       this.customRootConfig = null;
     }
     this.initializeRootConfig();
-    this.config = this.mergeConfig(this.localProjectConfigPath);
+    this.config = this.mergeConfig();
   }
 
   getLocalProjectConfigFile() {
@@ -40,7 +41,7 @@ class TemplateConfig {
   reset() {
     eleventyConfig.reset();
     this.initializeRootConfig();
-    this.config = this.mergeConfig(this.localProjectConfigPath);
+    this.config = this.mergeConfig();
   }
 
   getConfig() {
@@ -61,7 +62,6 @@ class TemplateConfig {
 
   initializeRootConfig() {
     this.rootConfig = this.customRootConfig || require("./defaultConfig.js");
-
     if (typeof this.rootConfig === "function") {
       this.rootConfig = this.rootConfig(eleventyConfig);
       // debug( "rootConfig is a function, after calling, eleventyConfig is %o", eleventyConfig );
@@ -69,11 +69,39 @@ class TemplateConfig {
     debug("rootConfig %o", this.rootConfig);
   }
 
-  mergeConfig(localProjectConfigPath) {
+  processPlugins(localConfig) {
+    // eleventyConfig.plugins
+    eleventyConfig.plugins.forEach(({ plugin, options }) => {
+      eleventyConfig.dir = localConfig.dir;
+      // TODO support function.name in plugin config functions
+      debug("Adding plugin (unknown name: check your config file).");
+      let pluginBench = aggregateBench.get("Configuration addPlugin");
+      if (typeof plugin === "function") {
+        pluginBench.before();
+        let configFunction = plugin;
+        configFunction(eleventyConfig, options);
+        pluginBench.after();
+      } else if (plugin && plugin.configFunction) {
+        pluginBench.before();
+        if (options && typeof options.init === "function") {
+          options.init.call(eleventyConfig, plugin.initArguments || {});
+        }
+
+        plugin.configFunction(eleventyConfig, options);
+        pluginBench.after();
+      } else {
+        throw new UserConfigError(
+          "Invalid EleventyConfig.addPlugin signature. Should be a function or a valid Eleventy plugin object."
+        );
+      }
+    });
+  }
+
+  mergeConfig() {
     let localConfig = {};
     let path = TemplatePath.join(
       TemplatePath.getWorkingDir(),
-      localProjectConfigPath
+      this.localProjectConfigPath
     );
     debug(`Merging config with ${path}`);
 
@@ -86,7 +114,6 @@ class TemplateConfig {
 
         localConfig = require(path);
         // debug( "localConfig require return value: %o", localConfig );
-
         if (typeof localConfig === "function") {
           localConfig = localConfig(eleventyConfig);
           // debug( "localConfig is a function, after calling, eleventyConfig is %o", eleventyConfig );
@@ -113,11 +140,17 @@ class TemplateConfig {
       debug("Eleventy local project config file not found, skipping.");
     }
 
+    // Delay processing plugins until after the result of localConfig is returned
+    // But BEFORE the rest of the config options are merged
+    // this way we can pass directories and other template information to plugins
+    this.processPlugins(localConfig);
+
     let eleventyConfigApiMergingObject = eleventyConfig.getMergingConfigObject();
 
     // remove special merge keys from object
+
     let savedForSpecialMerge = {
-      templateFormatsAdded: eleventyConfigApiMergingObject.templateFormatsAdded
+      templateFormatsAdded: eleventyConfigApiMergingObject.templateFormatsAdded,
     };
     delete eleventyConfigApiMergingObject.templateFormatsAdded;
