+++ src/UserConfig.js
@@ -4,7 +4,6 @@ const semver = require("semver");
 const { DateTime } = require("luxon");
 const EleventyBaseError = require("./EleventyBaseError");
 const bench = require("./BenchmarkManager").get("Configuration");
-const aggregateBench = require("./BenchmarkManager").get("Aggregate");
 const debug = require("debug")("Eleventy:UserConfig");
 const pkg = require("../package.json");
 
@@ -64,6 +63,8 @@ class UserConfig {
     this.dataExtensions = new Map();
 
     this.quietMode = false;
+
+    this.plugins = [];
   }
 
   versionCheck(expected) {
@@ -264,27 +265,7 @@ class UserConfig {
   }
 
   addPlugin(plugin, options) {
-    // TODO support function.name in plugin config functions
-    debug("Adding plugin (unknown name: check your config file).");
-    let pluginBench = aggregateBench.get("Configuration addPlugin");
-    if (typeof plugin === "function") {
-      pluginBench.before();
-      let configFunction = plugin;
-      configFunction(this, options);
-      pluginBench.after();
-    } else if (plugin && plugin.configFunction) {
-      pluginBench.before();
-      if (options && typeof options.init === "function") {
-        options.init.call(this, plugin.initArguments || {});
-      }
-
-      plugin.configFunction(this, options);
-      pluginBench.after();
-    } else {
-      throw new UserConfigError(
-        "Invalid EleventyConfig.addPlugin signature. Should be a function or a valid Eleventy plugin object."
-      );
-    }
+    this.plugins.push({ plugin, options });
   }
 
   getNamespacedName(name) {
@@ -325,7 +306,9 @@ class UserConfig {
 
   _normalizeTemplateFormats(templateFormats) {
     if (typeof templateFormats === "string") {
-      templateFormats = templateFormats.split(",").map(format => format.trim());
+      templateFormats = templateFormats
+        .split(",")
+        .map((format) => format.trim());
     }
     return templateFormats;
   }
@@ -629,7 +612,7 @@ class UserConfig {
       Object.assign(
         {
           key: fileExtension,
-          extension: fileExtension
+          extension: fileExtension,
         },
         options
       )
@@ -680,7 +663,7 @@ class UserConfig {
       dataExtensions: this.dataExtensions,
       extensionMap: this.extensionMap,
       quietMode: this.quietMode,
-      events: this.events
+      events: this.events,
     };
   }
 }
